embeddings/scripts/requirements.txt
transformers
umap-learn
sklearn
torch
click